Add frontend and backend prototypes
Implement a very simple datastore and frontend as a proof of concept.  Stub in or
naively implement gib methods so it will compile.  simpledb stores all posts in memory
and makes no effort to prevent starvation.  simplefrontend allows for posting replies and
starting new threads by setting the replyto field to -1.  It does no input validation or
escaping.